Main

Description

Este módulo define funções comuns da Tarefa 5 do trabalho prático.

O objetivo desta tarefa é juntar as funções criadas em tarefas anteriores e criar uma interface gráfica para fazer um jogo.

Desafio : Escolher o tema e arranjar imagens para o jogo.

Como tema do nosso jogo escolhemos o espaço, cada jogador é uma nave espacial e os Blocos Destrutíves são asteroides e os Blocos Indestrutíveis são pedras Impenetráveis

A maior partes as imagens são desenhadas com um editor de imagem (Gimp) à excessão de alguns elementos

Desafio : Como adicionar vários menus

Para conseguirmos cumprir este desafio aumentamos o nosso estado de modo a que possuísse um número que representasse o estado

Entre 0 e 1 -> Menu Inicial

Entre 1 e 2 -> Jogo em andamento

Entre 2 e 3 -> Jogo em Pausa

Entre 3 e 4 -> Jogo Terminou

Os pontos acima indicam o que cada numero representa

Desafio : Como fazer um contador de tempo

Para conseguirmos cumprir este desafio aumentamos o nosso estado de modo a que recebesse um inteiro a qual era adicionado uma unidade sempre que passava um tick. Como a framerate do nosso jogo é 4 frames por segundo é necessário dividir este número por quatro antes de ser desenhado.

Desafio : Opções de jogo.

O jogador pode escolher jogar com outros amigos no mesmo PC (PvP) ou contra 3 bots programados na tarefa 6. Também pode escolher de entre 5 mapas, reiniciar ou para jogo. O jogo acaba assim que passar um minuto desde o início do jogo ou se só restar apenas um ou nenhum jogador.

Em conclusão, a nossa interface gráfica é simples mas de fácil interpretação.

Documentation

type EstadoMain = (Estado, Float, Int, Int, [Picture])

Função principal da Tarefa 5.

NB: Esta Tarefa é completamente livre. Deve utilizar a biblioteca gloss para animar o jogo, e reutilizar de forma completa as funções das tarefas anteriores. | Este módulo define funções comuns da Tarefa 5 do trabalho prático.

Função principal da Tarefa 5.

NB: Esta Tarefa é completamente livre. Deve utilizar a biblioteca gloss para animar o jogo, e reutilizar de forma completa as funções das tarefas anteriores.

loadIMG :: IO [Picture]

Função para carregar as imagens

disM :: Display

Cria o ecrã do jogo

estadoInicial :: [Picture] -> EstadoMain

Dá o estado inicial do jogo

desenhaEstado :: EstadoMain -> Picture

Desenha o jogo dependendo do seu EstadoMain

drawall :: [Picture] -> [Picture] -> [Picture] -> [Picture] -> [Picture]

Junta um conjunto de figuras

escolhechoque :: Int -> [Picture] -> Picture

Escolhe a imagem do choque a usar dependendo do jogador que o usou

escolhecanhao :: Int -> Direcao -> [Picture] -> Picture

Escolhe a imagem do canhão a usar dependendo da sua direção e do jogador que o usou

escolhejogador :: Int -> Direcao -> [Picture] -> Picture

Escolhe a imagem do jogador a usar dependendo do seu indice na lista de jogadores e da sua direção

escolhelaser :: Int -> Direcao -> [Picture] -> Picture

Escolhe a imagem do laser a usar dependendo da sua direção e do jogador que o usou

drawscore :: [Picture] -> [Picture] -> [Picture] -> Picture

drawnumboflives :: EstadoMain -> Int -> [Picture]

Desenha o número de vidas dos Jogadores

drawnumboflazers :: EstadoMain -> Int -> [Picture]

Desenha o número de lasers dos Jogadores

drawnumbofchoque :: EstadoMain -> Int -> [Picture]

Desenha o número de choques dos Jogadores

drawmap :: EstadoMain -> Posicao -> [Picture]

Desenha o Mapa

drawshots :: EstadoMain -> [Picture]

Desenha os DisparoCanhao e DiparoChoque

drawlasers :: EstadoMain -> [[Picture]]

Desenha todos os DisparoLaser

drawonelaser :: Int -> Direcao -> [PosicaoGrelha] -> [Picture] -> [Picture]

Desenha um DisparoLaser

posicoesLaser :: PosicaoGrelha -> Direcao -> Mapa -> [PosicaoGrelha]

Encontra as posições afetadas por um DisparoLaser

drawplayers :: EstadoMain -> Int -> [Picture]

Desenha os Joagadores

eventChange :: Event -> EstadoMain -> EstadoMain

Altera o EstadoMain de acordo com a tecla pressionada e o o tipo de estado atual

bot1 :: Estado -> Estado

Aplica o bot da Tarefa6 ao jogador de indice 1

bot2 :: Estado -> Estado

Aplica o bot da Tarefa6 ao jogador de indice 2

bot3 :: Estado -> Estado

Aplica o bot da Tarefa6 ao jogador de indice 3

endgame :: EstadoMain -> EstadoMain

Acaba com um jogo (se chegar a um certo tempo ou se só ouver um jogador ou menos no estado)

winner :: [Jogador] -> (Int, Int, Int) -> Float

altera o EstadoMain para a apresentar o vencedor do jogo

onlyoneplayer :: [Jogador] -> Int -> Bool

Verifica se só há um ou zero jogadores no Estado

timeChange :: Float -> EstadoMain -> EstadoMain

Alera o Estado a cada frame

main :: IO ()

Cria o jogo