Tarefa3_2018li1g077

Description

O objetivo desta tarefa é comprimir um estado de jogo o melhor possivel

Desafio : Comprimir um mapa usando o sistema binário.

Um mapa pode ser composto por 3 pessas diferentes (Bloco Destrutivel,Bloco Indestrutivel e Vazia) podemos associar a cada uma destas peças valores binários (funcão tornaBinário). Tendo em conta que usando Haskell temos à disposição cerca de 1100000 de carateres podemos usar no máximo 20 bits para representar um carater, porque 2^21 (número de combinações binárias possiveis com 21 bits) já excede o número de carateres disponíveis. Como cada peça do mapa precisa de 2 bits para ser representada conseguimos ter uma média de compressão de 9.5 peças por caracter Um dos problemas que tivemos foi ter de usar um bit para marcar o fim de um caracter para conseguirmos fazer a descompressão corretamente por isso temos uma média de compressão de pessa de 9.5 por carater e não de 10.

Desafio : Comprimir a lista de jogadores e a lista de disparos

Para além de querermos ter uma boa compressão também queriamos que a nossa função fosse capaz de comprimir um grande número de jogadores com varias vidas e disparos. Para conseguirmos fazer isto criamos uma função que calcula o número de bits necessários para comprimir algo (se receber um lista de jogadores com jogadores que tem como o número máximo de vida 3 dá-nos que é necessário 2 bits para comprimir as vidas mas se fossem 4 já ia dar que eram necessários 3 bits) Deste modo conseguimos comprimir mapas de todos os tipos

Em conclusão esta tarefa foi feita de modo a desefiarmo-nos e deu-nos bastante informação sobre o sistema binário.

Testes

testesT3 :: [Estado]

Testes unitários da Tarefa 3.-

Cada teste é um Estado.

Funções principais da Tarefa 3.

comprime :: Estado -> [Char]

Comprime um Estado para formato textual.

NB: A função show representa um Estado num formato textual facilmente legível mas extenso.

NB: Uma boa solução deve representar o Estado dado no mínimo número de caracteres possível.

tiraNumDisparos :: [Disparo] -> [[Int]]

dPStringN2 :: Direcao -> Int

projectilParaString :: [Disparo] -> String

parParaString :: (Int, Int) -> String

dPString :: Direcao -> String

tiraPBala :: Disparo -> Posicao

tiraInBala :: Disparo -> Int

tiraDirBala :: [Disparo] -> [Direcao]

disparoParaChar :: [Disparo] -> String

disPAux :: Disparo -> [Int]

disparosC :: [Disparo] -> [Disparo]

disparosL :: [Disparo] -> [Disparo]

tiraVLC :: Jogador -> [Int]

maxAttribute :: [Jogador] -> Int

vlcParaBin :: Int -> [Jogador] -> String

tiraDJ :: Jogador -> Direcao

aux :: [a] -> [a]

checkBits :: Int -> Int -> Int -> Int

listPosTripl :: [Posicao] -> Int -> [(Int, Int, Int)]

ordena :: [(Int, Int, Int)] -> [(Int, Int, Int)]

ordenaAux :: (Int, Int, Int) -> [(Int, Int, Int)] -> [(Int, Int, Int)]

indToBin :: Int -> [Int] -> [String]

tttr :: (Int, Int, Int) -> Int

posicoesJogador :: Int -> Int -> String

addPosMap :: [Int] -> String -> String

posParaIndice :: [Posicao] -> Int -> [Int]

tira15 :: String -> String

mapaBin :: [[Peca]] -> String

tornaBinario :: Peca -> String

tornaBinarioD :: Direcao -> String

contaPecas :: String -> Int

contaPecasAux :: Int -> String -> Int

binParaInt :: String -> Int

descomprime :: String -> Estado

Descomprime um Estado no formato textual utilizado pela função comprime.

NB: A função comprime é válida de for possível recuperar o Estado utilizando a função descomprime, i.e.:

descomprime . comprime = id

NB: Esta propriedade é particularmente válida para a solução pré-definida:

read . show = id

a :: [Char] -> [Int]

funções auxiliares da função "descomprime"

dPString2n :: (Num a, Eq a) => a -> Direcao

Converte um número na sua posição correspondente

lNumToD :: [Int] -> [Disparo]

Transforma uma lista de numeros numa lista de disparos

lNumToDa :: [Int] -> Disparo

função auxiliar da função "lNumToD". A partir de uma lista de 5 números, cria um único disparo.

lNumToDa [0,1,2,3,4] = DisparoCanhao (1,2) 3 C
lNumToDa [1,3,9,1,1] = DisparoLaser (3,9) 1 E

dPString2 :: Char -> Direcao

Converte um char na sua direçao correspondente

charToChq :: String -> [Disparo]

Partindo duma string, devolve uma lista de disparos do tipo DisparoChoque

charToChqA :: [Int] -> [Disparo]

Função auxiliar da função charToChq. Percorre a lista de numeros gerada na função charToChq, recolhendo 2 valores (Int, Tick) correspondentes a cada DisparoChoque

criaLJogador

Arguments

:: [Posicao]
Lista das posicoes organizadas por ordem crescente dos indices dos jogadores

-> [Direcao]
Lista das direções de cada jogador, organizadas por ordem crescente dos indices

-> [Int]
Lista de numeros que correspondem ás vidas, lasers e choques dos jogadores.

-> [Jogador]

Cria a lista de jogadores

prepStringD :: String -> String

Prepara uma string de chars numa de 0's e 1's , pronta a ser lida

charParaLD :: String -> [Direcao]

Associa lista de 1's e 0's a direcoes

posicaoJogadores :: [Int] -> [Posicao] -> [(Int, Int, Int)]

Funcao que associa uma lista de indices a uma lista de posicoes

ordenaPosJog [2,0,1] [(1,3),(1,4),(2,2)] = [(1,3,2),(1,4,0),(2,2,1)]

ordenaPosJog :: [(Int, Int, Int)] -> [(Int, Int, Int)]

Funcao que ordena uma lista de triplos pelo indice do jogador

ordenaPosJog [(1,3,2),(1,4,0),(2,2,1)] = [(1,4,0),(2,2,1),(1,3,2)] 

ordPJA :: (Int, Int, Int) -> [(Int, Int, Int)] -> [(Int, Int, Int)]

Função auxiliar da função ordenaPosJog, insere um tuple no sitio adequado da lista.

doisP :: (Int, Int, Int) -> (Int, Int)

Funcao que tira o ultimo elemento duma tripla.

charParaMap :: String -> Mapa

charParaInd :: String -> [Int]

Sabendo o número de bits usados para representar os números (ord x), transforma uma string numa lista de números

binParaPos :: String -> [Posicao]

bit8Num :: Int -> String -> [Int]

criaMapa :: Char -> [Peca] -> Mapa

indParaPos :: Int -> Posicao -> Int -> Posicao

criaPos :: Int -> String -> [Int] -> [Int]

removeJogadores :: String -> String

remove os 11's da string de 1's e 0's

aux2 :: [Peca] -> [Peca]

agrupaBin :: Int -> String -> [String]

Agrupa uma string numa lista com strings de comprimento indicado

agrupaListas :: String -> [Peca]

agrupaListasAux :: String -> Peca

pegaLinha :: String -> Int -> String

Transforma um numero numa string de 0's e 1's

aux3 :: Int -> String