Tarefa4_2018li1g077

Description

Este módulo define funções comuns da Tarefa 4 do trabalho prático.

Testes

testesT4 :: [Estado]

Testes unitários da Tarefa 4.

Cada teste é um Estado.

Funções principais da Tarefa 4.

tick

Arguments

:: Estado
O Estado anterior.

-> Estado
O Estado após um Tick.

Avança o Estado do jogo um Tick de tempo.

NB: Apenas os Disparos afetam o Estado do jogo com o passar do tempo.

NB: Deve chamar as funções tickChoques, tickCanhoes e tickLasers pela ordem definida.

tickLasers :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos dos tiros de Laser disparados.

juntaXpl :: [([PosicaoGrelha], [Posicao])] -> ([PosicaoGrelha], [Posicao]) -> ([PosicaoGrelha], [Posicao])

Junta uma lista de Posicoes afetadas pela explosao

explJ :: PosicaoGrelha -> Direcao -> [Jogador] -> ([PosicaoGrelha], [Posicao])

Sabendo a posicaoGrelha que afeta um jogador, devolve a posicaoGrelha ocupada pelo jogador

expl :: Posicao -> Direcao -> ([PosicaoGrelha], [Posicao])

Recebe uma posicãoGrelha e uma direção, devolve as posições do mapa e da grelha que deverão ser afetadas pela explosão.

expL :: PosicaoGrelha -> Direcao -> ([PosicaoGrelha], [Posicao])

Sabendo uma posicão da grelha e uma direção, devolve as posicoes associadas aos 2 blocos á frente dessa posição.

efeitosExpl :: Estado -> ([PosicaoGrelha], [Posicao]) -> Estado

Sabendo as posições da grelha e do mapa afetados por uma explosão, muda o estado.

boom :: [Posicao] -> Mapa -> Mapa

Sabendo as posições do mapa a arrebentar, arrebenta com os blocos destrutíveis

boomJg :: [PosicaoGrelha] -> [Jogador] -> [Jogador]

Sabendo as posições da grelha afetadas por uma explosão, aplica os efeitos das explosões nos jogadores (acumulador)

boomDs :: [PosicaoGrelha] -> [Disparo] -> ([Disparo], [PosicaoGrelha])

Sabendo as posições da grelha afetadas por uma explosão, aplica os efeitos das explosões nos disparosCanhao.

laser

Arguments

:: Disparo
DisparoLaser a ser analisado

-> Mapa
Mapa do jogo.

-> [Jogador]
Lista de jogadores

-> [Disparo]
Lista de disparos

-> ([PosicaoGrelha], [Posicao])
Acumulador

-> ([PosicaoGrelha], [Posicao])

verificaDisparos :: PosicaoGrelha -> [Disparo] -> Bool

verifica se existe algum disparo no caminho do projéctil, e se existir, remove-o da lista de disparos (laser)

verificaJogadores :: PosicaoGrelha -> Direcao -> [Jogador] -> Bool

Verifica se um jogador é atingido pelo disparo, conhecendo a sua posicão e direcão.

checkJg :: PosicaoGrelha -> Direcao -> [PosicaoGrelha]

Transforma uma posicaoGrelha numa lista de posicões. Usada pela funcao verificaJogadores para verificar se um Disparo atingiu um jogador.

compLista :: Eq a => [a] -> [a] -> Bool

Verifica se existe algum elementa da primeira lista que pertence á segunda.

pdContinuarL :: (Peca, Peca) -> (Bool, Bool)

Verifica se o DisparoLaser pode continuar no seu caminho, dependendo do que há á sua frente, e decidindo se há explosão ou não). O primeiro elemento do par decide se o projéctil poderá continuar, o segundo decide se deverá criar uma explosão.

pdContinuarC :: (Peca, Peca) -> (Bool, Bool)

Verifica se o DisparoCanhao pode continuar no seu caminho, dependendo do que há á sua frente, e decidindo se há explosão ou não). O primeiro elemento do par decide se o projéctil poderá continuar, o segundo decide se deverá criar uma explosão.

checkB :: Mapa -> Posicao -> Direcao -> (Peca, Peca)

Verifica os blocos que se encontram junto do disparo, dependendo a direção

tickCanhoes :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos das balas de Canhao disparadas.

tiraPGP :: ([PosicaoGrelha], [Posicao], [Disparo]) -> ([PosicaoGrelha], [Posicao])

tiraD :: ([PosicaoGrelha], [Posicao], [Disparo]) -> [Disparo]

rmvD :: [Disparo] -> [Disparo]

Remove os disparos que "colidiram" sem ocupar a mesma posição (com acumulador)

disp :: Disparo -> (PosicaoGrelha, Direcao)

Recebe um disparo, faz um par que contêm a sua posicão e direção.

compPGD :: (PosicaoGrelha, Direcao) -> (PosicaoGrelha, Direcao) -> (PosicaoGrelha, Direcao) -> Bool

Compara pares com uma posicaoGrelha e direçao

dirOposta :: Direcao -> Direcao

Devolve a direção oposta á direção recebida.

rmPosD :: [Disparo] -> [Disparo]

Remove os disparos com a mesma posição

deLete :: Eq a => a -> [a] -> [a]

Remove todas as instâncias de um certo elemento de uma lista

canhao

Arguments

:: Disparo
DisparoLaser a ser analisado

-> Mapa
Mapa do jogo.

-> [Jogador]
Lista de jogadores

-> [Disparo]
Lista de disparos

-> ([PosicaoGrelha], [Posicao])
Acumulador

-> ([PosicaoGrelha], [Posicao], [Disparo])

Posições onde deve haver uma explosão causada por um disparoCanhao

tickChoques :: Estado -> Estado

Avança o Estado do jogo um Tick de tempo, considerando apenas os efeitos dos campos de Choque disparados.

efeitoChoques :: [Disparo] -> [Disparo]

A remove um tick a cada DisparoChoque, remove o DisparoChoque se tiver 0 Ticks restantes.

disparosC :: [Disparo] -> [Disparo]

Seleciona os DisparoChoque duma lista de disparos.

disparosL :: [Disparo] -> [Disparo]

Seleciona os DisparoCanhao e DisparoLaser duma lista de disparos.

disparosLser :: [Disparo] -> [Disparo]

Seleciona os DisparoLaser duma lista de disparos.

disparosCa :: [Disparo] -> [Disparo]

Seleciona os DisparoCanhao duma lista de disparos.